refactor: move EmailInUseError to Domain
