+++ src/domain/errors/account/email-in-use-error.ts
