+++ src/domain/errors/account/index.ts
